Amended any (Integer) criteriaUniqueResult() to Long(s) (avoids class cas exceptions if/when Hibernate is upgraded to later version)
git-svn-id: svn://localhost/ARK@6464 35ee9b83-dd2c-47f4-99a2-c706670c48ce

+++ ark-lims/src/main/java/au/org/theark/lims/model/dao/BioTransactionDao.java
@@ -66,8 +66,8 @@ public class BioTransactionDao extends HibernateSessionDao implements IBioTransa
 		Criteria criteria = buildBioTransactionCriteria(bioTransaction);
 		criteria.setProjection(Projections.rowCount());
 		
-		Integer totalCount = (Integer) criteria.uniqueResult();
-		return totalCount;
+		Long totalCount = (Long) criteria.uniqueResult();
+		return totalCount.intValue();
 	}
 
 	public List<BioTransaction> searchPageableBioTransactions(BioTransaction bioTransaction, int first, int count) {
